refactor : 데이터 베이스에서 발생한 예외 처리 Advice 구현
