+++ src/main/java/wooteco/subway/admin/common/controller/IndexController.java
@@ -1,4 +1,4 @@
-package wooteco.subway.admin.common.advice;
+package wooteco.subway.admin.common.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
